Begin modular refactor
All code is in no state to be able to be tested, unit or otherwise.

Basic python standards have not been followed, such that some modules
simply can not be tested due to acting as their own cgi scripts, or
isn't following PEP standards appropiately in such a way that putting
this in package format would be detrimental to each and every script
because of import issues.

I've started refactoring however I can't even tell if some of the code
is used or not. Regardless the majority is not production or even
development ready.

No testing can be done until refactoring is complete.